+++ src/components/forms/contact-us/contact-us.md
@@ -1,0 +1,3 @@
+```js
+<ContactUsForm />
+```
