+++ src/components/forms/index.js
@@ -1,0 +1,2 @@
+export { default as ContactUsForm } from './contact-us';
+export { default as Input } from './components/input';
