+++ src/components/modals/index.js
@@ -1,0 +1,2 @@
+export { default as Modal } from './modal';
+export { default as ContactUsModal } from './contact-us';
