+++ src/components/modals/modal/styles.scss
